Trivially extract StoredFieldsWriter from LucenePrimaryKeySegmentIndexV2
